Refactoring energy modules into failure generators
