+++ MobileNode.ned
@@ -7,7 +7,7 @@ import inet.environment.common.PhysicalEnvironment;
 import inet.mobility.contract.IMobility;
 
 import projeto.communication.protocols.base.IProtocol;
-import projeto.mobility.energy.base.IMobilityEnergy;
+import projeto.mobility.failures.base.IFailureGenerator;
 
 module MobileNode extends AdhocHost
 {   
@@ -15,6 +15,7 @@ module MobileNode extends AdhocHost
   		int internalMobNodeId = -1;
   		int simulationIndexOfFirstNode = 6;
   		int nodeType = 171;
+  		int numFailures = default(0);
   		
   		
   		@class(MobileNode);
@@ -22,11 +23,16 @@ module MobileNode extends AdhocHost
   	submodules:
   	    protocol: <default("CommunicationProtocolBase")> like IProtocol {
   	    }
-  	    energy: <default("MobilityEnergyBase")> like IMobilityEnergy {
+  	    failures[numFailures]: <default("FailureGeneratorBase")> like IFailureGenerator {
   	    }
   	
   	connections:
   		app[0].protocolGate <--> protocol.communicationGate++;
   		protocol.mobilityGate++ <--> mobility.protocolGate++;
-  		energy.mobilityGate <--> mobility.protocolGate++;
+
+  		for i=0..numFailures-1 {
+      		failures[i].mobilityGate <--> mobility.protocolGate++;
+  		}
+  		
+  		
 }
