+++ mobility/failures/SimpleConsumptionEnergy.h
