+++ mobility/failures/SimpleTimerEnergy.cc
@@ -13,7 +13,8 @@
 // along with this program.  If not, see http://www.gnu.org/licenses/.
 // 
 
-#include "SimpleTimerEnergy.h"
+#include "../failures/SimpleTimerEnergy.h"
+
 #include "../../communication/messages/internal/MobilityCommand_m.h"
 
 namespace projeto {
