+++ mobility/failures/SimpleTimerEnergy.h
@@ -13,20 +13,36 @@
 // along with this program.  If not, see http://www.gnu.org/licenses/.
 // 
 
-#ifndef MOBILITY_ENERGY_BASE_MOBILITYENERGYBASE_H_
-#define MOBILITY_ENERGY_BASE_MOBILITYENERGYBASE_H_
+#ifndef MOBILITY_FAILURES_SIMPLETIMERENERGY_H_
+#define MOBILITY_FAILURES_SIMPLETIMERENERGY_H_
 
+#include "../../communication/messages/internal/Telemetry_m.h"
 #include <omnetpp.h>
+
 using namespace omnetpp;
 
 namespace projeto {
 
-class MobilityEnergyBase : public cSimpleModule {
+class SimpleTimerEnergy : public cSimpleModule {
+    public:
+        virtual void initialize(int stage) override;
+
+        virtual void handleMessage(cMessage *msg) override;
+
+        virtual ~SimpleTimerEnergy();
+
+    public:
+        simtime_t batteryRTLDuration;
+        simtime_t batteryShutdownDuration;
+        simtime_t idleDuration;
+
     protected:
-        void initialize(int stage) {};
-        void handleMessage(cMessage *msg) {};
+        Telemetry currentTelemetry;
+        cMessage *RTLMessage;
+        cMessage *shutdownMessage;
+        bool isReturning;
 };
 
 } /* namespace projeto */
 
-#endif /* MOBILITY_ENERGY_BASE_MOBILITYENERGYBASE_H_ */
+#endif /* MOBILITY_FAILURES_SIMPLETIMERENERGY_H_ */
