+++ mobilityDrones-omnetpp.ini
@@ -328,10 +328,12 @@ description="Simulation with one drone with wifi radio only"
 *.quads[*].app[*].timeoutDuration = 10s
 *.quads[0].mobility.startTime = 0s
 *.quads[0].app[*].startTime = normal(1s, 1s)
-*.quads[0].energy.typename = "SimpleTimerEnergy"
-*.quads[0].energy.batteryRTLDuration = 10s
-*.quads[0].energy.batteryShutdownDuration = 25s
-*.quads[0].energy.idleDuration = 5s
+*.quads[0].numFailures = 1
+*.quads[0].failures[0].typename = "SimpleConsumptionEnergy"
+*.quads[0].failures[0].batteryCapacity = 400mAh
+*.quads[0].failures[0].batteryRTLThreshold = 500mAh
+*.quads[0].failures[0].batteryConsumption = 1.5A
+*.quads[0].failures[0].idleDuration = 5s
 
 *.quads[0].energy.batteryConsumption = 1.5A
